Update GaiaHubmod.py
Changes to deal with low number of stars by adding a padding in the pmra and pmdec before the transpose. Also added a try and except in case it fails to determine the footprint of the chip which is currently happening for some MAST products. Likely a result of old fits standards or WCS info.